puyuan/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.5
Django==4.2.6
djangorestframework==3.14.0
idna==3.4
jsonpickle==3.0.2
jwt==1.3.1
pycparser==2.21
pytz==2023.3.post1
requests==2.31.0
Serialize==0.2.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
